apps/api/pyproject.toml
[project]
name = "dnd-campaign-map-api"
version = "0.1.0"
description = "Self-hosted FastAPI backend for a D&D campaign map UI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<1.0",
    "pydantic-settings>=2.2,<3.0",
    "uvicorn[standard]>=0.29,<1.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.29,<1.0",
    "alembic>=1.13,<2.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "itsdangerous>=2.1,<3.0",
    "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "anyio[trio]>=4.0,<5.0",
    "respx>=0.21,<1.0",
]

[project.scripts]
dndmap-db = "app.cli.db:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
asyncio_mode = "auto"

